move pipe_from_one_command_to_another test to its own suite
